pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "sec-bootstrapper"
version = "0.1.0"
description = "Modular Linux/Ubuntu security hardening bootstrapper"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "Security Team" }
]
keywords = ["security", "hardening", "ubuntu", "linux", "bootstrap"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: System :: Systems Administration",
]

dependencies = [
    "typer>=0.9.0",
    "pydantic>=2.0.0",
    "pydantic-yaml>=1.0.0",
    "rich>=13.0.0",
    "packaging>=23.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.0.0",
]

[project.scripts]
sec-bootstrapper = "sec_bootstrapper.cli.main:app"

[tool.hatch.build.targets.wheel]
packages = ["sec_bootstrapper"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "W", "C90"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"
